requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
django-filter==24.3
django-seed==0.3.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
Faker==35.0.0
importlib_metadata==8.5.0
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
zipp==3.21.0